Version 0.91
In previous versions, the Choose Name list would show incorrect names on searches if some of the names found had commas as in "last, first". This version fixes that issue.

In previous versions, entering text such as "","" could prevent that application from displaying properly. This version fixes that issue by doing a more proper job of creating and parsing CSV.

Although not visible in any way to users, I renamed all keys used in session storage so that they are guaranteed to be unique to this add-on, and I don't store CSV in session storage anymore.